# LS optimization: early stop, const arrays in moduleGapSize, simpler logic in isTighterTiltedModules 