- check inner module occupancy before looping over the segments: small speedup in one event test from 0.95 to 0.90 ms
- const arrays in moduleGapSize; provides most of the gain
-  logic in isTighterTiltedModules adds a few %

Summary:
- createSegmentsInGPUv2:
    - overall profiler for the first event: 0.76 ms -> 0.42 ms
    - register usage: 80 -> 72
- standard timing 
<img width="700" alt="image" src="https://github.com/SegmentLinking/TrackLooper/assets/4676718/6914face-a7e8-46ec-bd5f-a755b104bb3b">

There was someone else on the 8-head A100 node during the tests above; so, fluctuations are possible.