@slava77 2023-06-24
Summary using the nvidia cluster:
- createSegmentsInGPUv2:
    - overall profiler for the first event: 0.76 ms -> 0.42 ms
    - register usage: 80 -> 72
- standard timing 
<img width="700" alt="image" src="https://github.com/SegmentLinking/TrackLooper/assets/4676718/6914face-a7e8-46ec-bd5f-a755b104bb3b">

There was someone else on the 8-head A100 node during the tests above; so, fluctuations are possible.