@slava77 2023-06-30
@VourMa 
the memory use goes down indeed
<img width="1148" alt="image" src="https://github.com/SegmentLinking/TrackLooper/assets/4676718/89ac1abd-c3ff-4067-b77d-df6017270130">